add ExceptionHandler ; Add some custom exception

+++ src/main/java/com/shop/shop/service/impl/ProductServiceImpl.java
@@ -1,5 +1,6 @@
 package com.shop.shop.service.impl;
 
+import com.shop.shop.configuration.exception.product.ProductNotFoundException;
 import com.shop.shop.dto.CategoryDTO;
 import com.shop.shop.dto.ProductDTO;
 import com.shop.shop.entity.Category;
@@ -11,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -27,7 +29,12 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public Product getOne(Long id) {
-        return repository.getById(id);
+
+        Optional<Product> byId = repository.findById(id);
+        if (byId.isPresent()){
+            return byId.get();
+        }
+        throw new ProductNotFoundException("Product with id "+ id +" not found");
     }
 
     @Override
@@ -37,6 +44,7 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public List<Product> getAllByCategories(List<CategoryDTO> categories) {
+
         List<Category> collect = categories
                 .stream()
                 .map(categoryDTO -> categoryService.map(categoryDTO))
@@ -53,7 +61,6 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public Product update(ProductDTO productDTO) {
         if (productDTO.getId() == null){
-            // TODO: 04.11.2021 add custom exception
             throw new IllegalArgumentException("Id cant be null");
         }
         Product product = map(productDTO);
@@ -62,6 +69,9 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void delete(Long id) {
+        if (id == null){
+            throw new IllegalArgumentException("Id cant be null");
+        }
         repository.deleteById(id);
     }
